+++ src/wasm.h
@@ -783,6 +783,7 @@ public:
     WideIntMulId,
     WaitqueueNewId,
     WaitqueueNotifyId,
+    PublishId,
     NumExpressionIds
   };
   Id _id;
@@ -1838,6 +1839,16 @@ public:
 
   Expression* waitqueue;
   Expression* count;
+
+  void finalize();
+};
+
+class Publish : public SpecificExpression<Expression::PublishId> {
+public:
+  Publish() = default;
+  Publish(MixedArena& allocator) : Publish() {}
+
+  Expression* ref;
 
   void finalize();
 };