+++ src/wasm/wasm-validator.cpp
@@ -570,6 +570,7 @@ public:
   void visitStructWait(StructWait* curr);
   void visitWaitqueueNew(WaitqueueNew* curr);
   void visitWaitqueueNotify(WaitqueueNotify* curr);
+  void visitPublish(Publish* curr);
   void visitStringNew(StringNew* curr);
   void visitStringConst(StringConst* curr);
   void visitStringMeasure(StringMeasure* curr);
@@ -1244,6 +1245,11 @@ void FunctionValidator::visitLoad(Load* curr) {
                  curr,
                  "SIMD operations require SIMD [--enable-simd]");
   }
+  if (curr->type == Type::f32 && curr->bytes == 2) {
+    shouldBeTrue(getModule()->features.hasFP16(),
+                 curr,
+                 "FP16 operations require FP16 [--enable-fp16]");
+  }
   validateMemBytes(curr->bytes, curr->type, curr);
   validateOffset(curr->offset, memory, curr);
   validateAlignment(
@@ -1295,6 +1301,11 @@ void FunctionValidator::visitStore(Store* curr) {
     shouldBeTrue(getModule()->features.hasSIMD(),
                  curr,
                  "SIMD operations require SIMD [--enable-simd]");
+  }
+  if (curr->valueType == Type::f32 && curr->bytes == 2) {
+    shouldBeTrue(getModule()->features.hasFP16(),
+                 curr,
+                 "FP16 operations require FP16 [--enable-fp16]");
   }
   validateMemBytes(curr->bytes, curr->valueType, curr);
   validateOffset(curr->offset, memory, curr);
@@ -2378,6 +2389,13 @@ void FunctionValidator::visitUnary(Unary* curr) {
     case FloorVecF16x8:
     case TruncVecF16x8:
     case NearestVecF16x8:
+    case PromoteLowVecF16x8ToVecF32x4:
+    case DemoteZeroVecF32x4ToVecF16x8:
+    case DemoteZeroVecF64x2ToVecF16x8:
+    case TruncSatSVecF16x8ToVecI16x8:
+    case TruncSatUVecF16x8ToVecI16x8:
+    case ConvertSVecI16x8ToVecF16x8:
+    case ConvertUVecI16x8ToVecF16x8:
       shouldBeTrue(getModule()->features.hasFP16(),
                    curr,
                    "FP16 operations require FP16 [--enable-fp16]");
@@ -2432,17 +2450,10 @@ void FunctionValidator::visitUnary(Unary* curr) {
     case TruncSatZeroUVecF64x2ToVecI32x4:
     case DemoteZeroVecF64x2ToVecF32x4:
     case PromoteLowVecF32x4ToVecF64x2:
-    case PromoteLowVecF16x8ToVecF32x4:
-    case DemoteZeroVecF32x4ToVecF16x8:
-    case DemoteZeroVecF64x2ToVecF16x8:
     case RelaxedTruncSVecF32x4ToVecI32x4:
     case RelaxedTruncUVecF32x4ToVecI32x4:
     case RelaxedTruncZeroSVecF64x2ToVecI32x4:
     case RelaxedTruncZeroUVecF64x2ToVecI32x4:
-    case TruncSatSVecF16x8ToVecI16x8:
-    case TruncSatUVecF16x8ToVecI16x8:
-    case ConvertSVecI16x8ToVecF16x8:
-    case ConvertUVecI16x8ToVecF16x8:
       shouldBeEqual(curr->type, Type(Type::v128), curr, "expected v128 type");
       shouldBeEqual(
         curr->value->type, Type(Type::v128), curr, "expected v128 operand");
@@ -2617,7 +2628,7 @@ void FunctionValidator::visitRefAs(RefAs* curr) {
     case AnyConvertExtern: {
       shouldBeTrue(getModule()->features.hasGC(),
                    curr,
-                   "any.convert_extern requries GC [--enable-gc]");
+                   "any.convert_extern requires GC [--enable-gc]");
       if (curr->type == Type::unreachable) {
         return;
       }
@@ -2631,7 +2642,7 @@ void FunctionValidator::visitRefAs(RefAs* curr) {
     case ExternConvertAny: {
       shouldBeTrue(getModule()->features.hasGC(),
                    curr,
-                   "extern.convert_any requries GC [--enable-gc]");
+                   "extern.convert_any requires GC [--enable-gc]");
       if (curr->type == Type::unreachable) {
         return;
       }
@@ -3752,6 +3763,26 @@ void FunctionValidator::visitWaitqueueNotify(WaitqueueNotify* curr) {
                                     Type(Type::BasicType::i32),
                                     curr,
                                     "waitqueue.notify count must be an i32");
+}
+
+void FunctionValidator::visitPublish(Publish* curr) {
+  shouldBeTrue(
+    !getModule() || getModule()->features.hasSharedEverything(),
+    curr,
+    "publish requires shared-everything [--enable-shared-everything]");
+
+  shouldBeTrue(curr->ref->type == Type::unreachable || curr->ref->type.isRef(),
+               curr->ref,
+               "publish's argument should be a reference type");
+
+  if (curr->ref->type == Type::unreachable) {
+    shouldBeEqual(curr->type,
+                  Type(Type::unreachable),
+                  curr,
+                  "unreachable publish value must have unreachable type");
+  } else {
+    shouldBeEqual(curr->ref->type, curr->type, curr, "bad publish type");
+  }
 }
 
 void FunctionValidator::visitArrayNew(ArrayNew* curr) {
@@ -5160,7 +5191,7 @@ void validateGlobals(Module& module, ValidationInfo& info) {
     }
     FunctionValidator(module, &info).validate(curr->init);
     // If GC is enabled (which means globals can refer to other non-imported
-    // globals), check that globals only refer to preceeding globals.
+    // globals), check that globals only refer to preceding globals.
     if (module.features.hasGC() && curr->init) {
       for (auto* get : FindAll<GlobalGet>(curr->init).list) {
         auto* global = module.getGlobalOrNull(get->name);