+++ ts/test/binaryen.test.ts
@@ -185,7 +185,7 @@ suite("binaryen", () => {
 
 	test(".ExpressionId", () => {
 		// NOTE: the length is twice the number of members due to how TypeScript emits enums.
-		assert.strictEqual(Object.entries(binaryen.ExpressionId).length, 88 * 2);
+		assert.strictEqual(Object.entries(binaryen.ExpressionId).length, 89 * 2);
 
 		assert.strictEqual(binaryen.ExpressionId.Invalid, 0);
 		assert.strictEqual(binaryen.ExpressionId.Block, 1);
@@ -306,12 +306,13 @@ suite("binaryen", () => {
 		assert.strictEqual(binaryen.ExpressionId.WideIntMul, 106);
 		assert.strictEqual(binaryen.ExpressionId.WaitqueueNew, 107);
 		assert.strictEqual(binaryen.ExpressionId.WaitqueueNotify, 108);
+		assert.strictEqual(binaryen.ExpressionId.Publish, 109);
 	});
 
 
 	test(".SideEffect", () => {
 		// NOTE: the length is twice the number of members due to how TypeScript emits enums.
-		assert.strictEqual(Object.entries(binaryen.SideEffect).length, 17 * 2);
+		assert.strictEqual(Object.entries(binaryen.SideEffect).length, 18 * 2);
 
 		assert.strictEqual(binaryen.SideEffect.None, 0);
 		assert.strictEqual(binaryen.SideEffect.Branches, 1 << 0);
@@ -329,7 +330,8 @@ suite("binaryen", () => {
 		assert.strictEqual(binaryen.SideEffect.Throws, 1 << 12);
 		assert.strictEqual(binaryen.SideEffect.DanglingPop, 1 << 13);
 		assert.strictEqual(binaryen.SideEffect.TrapsNeverHappen, 1 << 14);
-		assert.strictEqual(binaryen.SideEffect.Any, (1 << 15) - 1);
+		assert.strictEqual(binaryen.SideEffect.Suspends, 1 << 15);
+		assert.strictEqual(binaryen.SideEffect.Any, (1 << 16) - 1);
 	});